Removed save() in favor of link()
Seemed silly to force a pending save at the end of a chain as the use
case usually won't use chaning all that much. The gain for ending a
chain like that is miniscual for the average use case of this class.
Chaining will still work but is no longer a requirement from the user's
stand point.